DL-31: Provide flag to disable zk based distributed lock
DL doesn't enforce any leader election. However it still provides a zookeeper ephemeral znode based lock for leader election. It is unnecessary if applications use core library directly already have its own leader election mechanism.

This change is to provide a flag to allow disable the zk based lock.

Author: Sijie Guo <sijieg@twitter.com>

Reviewers: Leigh Stewart <lstewart@apache.org>

Closes #9 from sijie/sijie/flag_to_disable_lock
